requirements.txt
opencv-python~=4.5.5.62
python-dotenv~=0.19.2
numpy~=1.21.5
websockets~=10.3